test: add node contract deployment test
