+++ test/node.js
@@ -1,0 +1,119 @@
+const { Wavelet, Contract, TAG_TRANSFER, JSBI, Buffer } = require(".."); //window["wavelet-client"];
+const util = require("util");
+const readFile = util.promisify(require("fs").readFile);
+const path = require("path");
+const BigInt = JSBI.BigInt;
+
+const client = new Wavelet("https://devnet.perlin.net");
+
+function toArrayBuffer(buffer) {
+  var ab = new ArrayBuffer(buffer.length);
+  var view = new Uint8Array(ab);
+  for (var i = 0; i < buffer.length; ++i) {
+      view[i] = buffer[i];
+  }
+  return ab;
+}
+
+(async () => {
+  console.log(Wavelet.generateNewWallet());
+  const wallet = Wavelet.loadWalletFromPrivateKey(
+    "87a6813c3b4cf534b6ae82db9b1409fa7dbd5c13dba5858970b56084c4a930eb400056ee68a7cc2695222df05ea76875bc27ec6e61e8e62317c336157019c405"
+  );
+
+  const contractCode = await readFile(path.resolve(__dirname, "./token_bg.wasm"));
+  console.log("Contract", contractCode);
+  const contractId = await client.deployContract(wallet, toArrayBuffer(contractCode), JSBI.BigInt(100000));
+  console.log("res", contractId);
+
+  console.log(await client.getNodeInfo());
+
+  const accountResponse = await client.getAccount(
+    "400056ee68a7cc2695222df05ea76875bc27ec6e61e8e62317c336157019c405"
+  );
+  console.log("account", accountResponse);
+  
+  try {
+    const txResponse = await client.transfer(
+      wallet,
+      "f03bb6f98c4dfd31f3d448c7ec79fa3eaa92250112ada43471812f4b1ace6467",
+      0
+    );
+    console.log("txResponse", txResponse);
+    await client.pollTransactions(
+      {
+        onTransactionApplied: data => {
+          console.log("tx applied", data);
+        }
+      },
+      {
+        id: txResponse.id
+      }
+    );
+
+    const transfer = await client.getTransaction(
+      txResponse.id
+    );
+    console.log(Wavelet.parseTransaction(transfer.tag, transfer.payload));
+  } catch (err) {
+    alert(err.message || err);
+  }
+
+
+  const account = await client.getAccount(
+    Buffer.from(wallet.publicKey).toString("hex")
+  );
+
+  console.log("account", account);
+  const contract = new Contract(
+    client,
+    "4b6b43eba9eb8ed7402e0b7103a3eab9dfba48be05e359557d8c71f6a8513563"
+  );
+  await contract.init();
+
+  console.log(
+    contract.test(wallet, "balance", BigInt(0), {
+      type: "raw",
+      value: "400056ee68a7cc2695222df05ea76875bc27ec6e61e8e62317c336157019c405"
+    })
+  );
+
+  console.log(
+    await contract.call(
+      wallet,
+      "balance",
+      BigInt(0),
+      JSBI.subtract(JSBI.BigInt(account.balance), JSBI.BigInt(1000000)),
+      1000n,
+      {
+        type: "raw",
+        value:
+          "400056ee68a7cc2695222df05ea76875bc27ec6e61e8e62317c336157019c405"
+      }
+    )
+  );
+
+  const consensusPoll = await client.pollConsensus({
+    onRoundEnded: console.log
+  });
+  const transactionsPoll = await client.pollTransactions(
+    { onTransactionApplied: console.log },
+    {
+      tag: TAG_TRANSFER,
+      creator:
+        "400056ee68a7cc2695222df05ea76875bc27ec6e61e8e62317c336157019c405"
+    }
+  );
+  const accountsPoll = await client.pollAccounts(
+    { onAccountUpdated: console.log },
+    { id: "400056ee68a7cc2695222df05ea76875bc27ec6e61e8e62317c336157019c405" }
+  );
+
+  for (let i = 0; i < 100; i++) {
+    await client.transfer(
+      wallet,
+      "e49e8be205a00edb45de8183a4374e362efc9a4da56dd7ba17e2dd780501e49f",
+      BigInt(1000000)
+    );
+  }
+})();
